NP-3386 Split files
